+++ scratch/test_deterministic_with_measurements.R
@@ -1,5 +1,5 @@
 
-# devtools::install_local( R'(C:\Users\james\OneDrive\Desktop\Git\dsem)', force = TRUE, dep = FALSE )
+#devtools::install_local( R'(C:\Users\James.Thorson\Desktop\Git\dsem)', force = TRUE, dep = FALSE )
 
 library(dsem)
 
@@ -12,55 +12,70 @@ make_ar = function(rho, X){
 # 
 # USE-CASE
 #  Determinstic and family = "fixed" with measurements
+# Only works with mvn_project
 #
 ################
 
-
-p_missing = 0.4
-
-X = cumsum( rnorm(100) )
-Y = 2 * X
-Z = 0.5*Y + rnorm(100)
-X_missing = sample(seq_along(X), size=p_missing*length(X), replace=FALSE)
-Y_missing = sample(seq_along(X), size=p_missing*length(X), replace=FALSE)
-Y_missing = union( setdiff( seq_along(X), X_missing ), Y_missing )
-
-# Missing-ness
-Xobs = X
-Xobs[X_missing] = NA
-Yobs = Y
-Yobs[Y_missing] = NA
-
-# Bundle
-dat = data.frame( 
-  X = 4 + Xobs, 
-  Y = 1 + Yobs, 
-  Z = 2 + Z 
-)
-
-library(dsem)
-
-sem = "
-  X -> X, 1, rho
-  X -> Y, 0, b_XY
-  Y -> Z, 0, b_YZ
-  Y <-> Y, 0, NA, 0
-"
-
-control = dsem_control(
-  gmrf_parameterization = "mvn_project",
-  project_k = is.na(dat)
-)
-fit = dsem(
-  tsdata = ts(dat),
-  sem = sem,
-  control = control
-)
+if( FALSE ){
+  set.seed(123)
+  p_missing = 0.4
+  
+  X = cumsum( rnorm(100) )
+  Y = 2 * X
+  Z = 0.5*Y + rnorm(100)
+  X_missing = sample(seq_along(X), size=p_missing*length(X), replace=FALSE)
+  Y_missing = sample(seq_along(X), size=p_missing*length(X), replace=FALSE)
+  Y_missing = union( setdiff( seq_along(X), X_missing ), Y_missing )
+  
+  # Missing-ness
+  Xobs = X
+  Xobs[X_missing] = NA
+  Yobs = Y
+  Yobs[Y_missing] = NA
+  
+  # Bundle
+  dat = data.frame( 
+    X = 4 + Xobs, 
+    Y = 1 + Yobs, 
+    Z = 2 + Z 
+  )
+  
+  library(dsem)
+  
+  sem = "
+    X -> X, 1, rho
+    X -> Y, 0, b_XY
+    Y -> Z, 0, b_YZ
+    Y <-> Y, 0, NA, 0
+  "
+  
+  control = dsem_control(
+    gmrf_parameterization = "mvn_project",
+    #gmrf_parameterization = "gmrf_project",
+    #project_k = is.na(dat),
+    build_model = TRUE,
+    nlminb_loops = 0,
+    getsd = FALSE,
+    extra = FALSE
+  )
+  fit = dsem(
+    tsdata = ts(dat),
+    sem = sem,
+    control = control
+  )
+  rep = fit$obj$report()
+  # RTMB::dgmrf( rep$dev_o, mu = rep(0,length(rep$dev_o)), Q = rep$Q_A, log = TRUE )
+  
+  rep$dev_u1
+  fit$tmb_inputs$data$unobs_idx
+}
 
 ################
 # 
 # USE-CASE
 #  logistic regression
+# runs and gives identical answer with either mvn_project or gmrf_project
+#  and gmrf_project is *much* faster
 #
 ################
 
@@ -84,9 +99,13 @@ sem = "
 
 # New option
 control = dsem_control(
-  gmrf_parameterization = "mvn_project",
+  gmrf_parameterization = c("gmrf_project", "mvn_project")[1],
   #build_model = FALSE,
-  use_REML = FALSE
+  use_REML = FALSE,
+  nlminb_loops = 1,
+  newton_loops = 0,
+  getsd = TRUE,
+  extra = TRUE
 )
 fit = dsem(
   tsdata = ts(dat),
@@ -94,12 +113,14 @@ fit = dsem(
   control = control,
   family = c("fixed", "bernoulli")
 )
+rep = fit$obj$report()
 
 # Old option
 sem = "
   X -> X, 1, rho
   X -> Y, 0, b_XY
   Y <-> Y, 0, NA, 0.0001
+  #Y -> X, 1, gamma
 "
 control = dsem_control(
   use_REML = FALSE